requirements.txt
pyserial>=3.4
numpy>=1.19.0
pygame>=2.0.0
matplotlib>=3.3.0
pyautogui>=0.9.50
bleak>=0.19.0
